[internal][TimePicker] Move useInterceptProps into module
